Change condition logic names
In PHP7, we could use `and()` and `or()` as method names. This would
make it harder to produce a PHP5 compatible version of this package.
The names `logicalAnd` and `logicalOr` are a bit wordy and the `where()`
name doesn't properly reflect the nature of conditions.

Using `with()`, `andWith()`, and `orWith()` feels more natural and reads
better, in the authors opinion.