(fix) Reset array keys for params()
When generating SQL, null values (and boolean values) are included in the SQL, and not included in the parameters.
By filtering out these parameters, their original index in the array is preserved, which leads a non-sequential keys. It seems PDOStatement::execute() isn't able to figure this out. 

Returning a new array with only the array values fixes this issue. 

HTH